// src/types/solar-facts.ts
export type SolarFacts = {
  area: number;
  calculatedElectricityMaxPower: number;
  calculatedElectricityModuleArea: number;
  calculatedThermalMaxPower: number;
  calculatedThermalModuleArea: number;
  electricityProduction: number;
  exposition: string;
  roofType: string;
  slope: string;
  suitability: string;
  thermalProduction: number;
};